refactor!: remove multirequest methods (#20)
Multirequest methods like `send_all` and `send_all_paginated` probably
make more sense as a user-implementation as there are conflicting
decisions in what the return type should be (single stream of results,
iterator of streams, etc). Single request queries are more unambiguous,
so we remove the multirequest feature and focus instead on great
single-request ergonomics, leaving the decision in how to send multiple
requests to the user.